Craigslist.py
This script is intended to post multiple ads on Craiglist with  similar
content. You can't post multiple ads on Craigslist when the content is
very similar for it will automatically remove them. With this Script,
you can have pre-made content that will be inserted into different ads
you will post. By having pre-made sentences still communicating what you
are promoting the changing of the sentences with this script will allow
you to post multiple ads as if they were different resulting the
automation Craigslist  to not remove your ads. It doesn't have a GUI,
but it is relatively easy to work with. I hope you find it useful :)
How to Use:
1.) The input file named craigslist.txt should be in the same folder as
the script. (I made this the default input file name but it can be
changed near the top of the script.)
2.) The beginning tag for the list of options is <select>. The end tag
for the list of options is </select>. Anything between a <select> and
</select> tag will be considered options for that selection. (see
craigslist.txt as an example)
3.) The separator between the options is a single bar | character by
itself.
4.) There should be NO bar at the start or the end of the options, only
in between the words or phrases that are choices for the script.
(see attached craigslist.txt as an example)
5.)The output of the script will be CraigslistFinished.txt in the same
folder as the script.
6.) I wrote it in Python using the formatting I'd learned for Python
3.4.1.
7.) I've tested the script with multiple lines of text and with phrase
options as well as single word options.