Automated File Transfer
A small coding assignment when I started using Python at The Tech
Academy (Portland Coding School). The task was to create a GUI using wx
python and have a method which will allow you to open a directory and
choose an origin folder and a destination folder to transfer files to
and from.
The project also uses sqlite3 and it stores the of time of execution of
the method and prints out the last time the code was ran from the
database.